Rework the API slurper and import the Clubhouse API
The slurper script has been refactored into two scripts, a slurper and a builder. The slurper now outputs its data to intermediate JSON files, and the builder consumes the files to generate Python. This approach allows us to trivialize any major architectural changes to the interface, and makes it easier to snake in future API revisions.

The v3 API has now been slurped and put into the Python client object.