Always explicitly build body dicts
I realized that the dict was including query parameters with the locals() call, so I altered the generator to explicitly build the dicts in those cases. I then realized that the size of the code didn't grow by an excessive amount if I just explicitly built all body parameter dicts.